Move some declarations to different files
