+++ src/core/heuristic.rs
@@ -1,4 +1,4 @@
-use super::{color::IconColor, MatchingState};
+use super::{Lang, MatchingState};
 use std::fmt;
 
 /// Trait for implementing heuristics to match directories and files for deletion.
@@ -25,72 +25,40 @@ impl fmt::Display for dyn Heuristic {
     }
 }
 
-/// Data structure representing a programming language or other reason for a match.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct Lang {
-    /// Name of the language or heuristic.
-    pub name: &'static str,
-    /// Icon representing the language or heuristic.
-    ///
-    /// This icon should be an emoji or [a nerd font symbol](https://www.nerdfonts.com/).
-    pub icon: &'static str,
-    /// Short name/abbreviation of the language or heuristic, used when icons are not supported.
-    pub short: &'static str,
-    /// [ANSI 8-bit color index](https://en.wikipedia.org/wiki/ANSI_escape_code#8-bit) for the language or heuristic.
-    pub color: IconColor,
-}
+/// Simplified heuristic declaration.
+///
+/// Parameters in order:
+/// - `name` - heuristic name, also used as a generated struct indentifier,
+/// - `icon` - Nerd Font icon (see [`Lang::icon`]),
+/// - `short` - heuristic name abbreviation (see [`Lang::short`]),
+/// - `color` - [`IconColor`] instance,
+/// - `state` - parameter name for [`Heuristic::check_for_matches()`],
+/// - `expression` - heuristic body with state in scope.
+#[macro_export]
+macro_rules! heuristic {
+    ($name:ident, $icon:expr, $short:literal, $color:expr, $state:ident, $expression:expr) => {
+        use $crate::core::{Heuristic, IconColor, Lang, MatchingState};
 
-impl PartialOrd for Lang {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
+        #[derive(Default)]
+        pub struct $name;
 
-impl Ord for Lang {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.name.cmp(other.name)
-    }
-}
+        #[allow(dead_code)]
+        pub const INSTANCE: $name = $name;
 
-impl fmt::Display for Lang {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.name)
-    }
-}
+        impl Heuristic for $name {
+            fn info(&self) -> &'static Lang {
+                const LANG: Lang = Lang {
+                    name: stringify!($name),
+                    icon: $icon,
+                    short: $short,
+                    color: $color,
+                };
+                &LANG
+            }
 
-/// Extended [`Lang`] data structure with a comment field.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CommentedLang {
-    /// Language this struct is based on.
-    pub lang: &'static Lang,
-    /// Additional comment.
-    pub comment: String,
-}
-
-impl fmt::Display for CommentedLang {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} - {}", self.lang.name, self.comment)
-    }
-}
-
-impl CommentedLang {
-    /// Returns the name of the language or heuristic.
-    #[inline]
-    pub fn name(&self) -> &str {
-        self.lang.name
-    }
-
-    /// Returns the icon representing the language or heuristic.
-    ///
-    /// This icon should be an emoji or [a nerd font symbol](https://www.nerdfonts.com/).
-    #[inline]
-    pub fn icon(&self) -> &str {
-        self.lang.icon
-    }
-
-    /// Returns the abbreviation of the language or heuristic, used when icons are not supported.
-    #[inline]
-    pub fn short(&self) -> &str {
-        self.lang.short
-    }
+            fn check_for_matches(&self, $state: &mut MatchingState) {
+                $expression
+            }
+        }
+    };
 }
